requirements.txt
PyQt6
psutil
pywin32
Pillow